Fix authentication module
